Java Farming Application
Have used different Swing components to implement an application with a GUI, practiced how to use event-driven programming, practiced to use streams to store the relevant data in external files, practiced JDBC API,Multithreading in Java and Object serialization API in Java, practiced how to use Java Security API (in particular, MD5 implementation).  Learned how to connect a Java application to a backend database; to learn how to use part of the object serialization and Java Security API, and multithreading.